% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_pheatmap.R
\name{marker_pheatmap}
\alias{marker_pheatmap}
\title{marker_pheatmap() Function}
\usage{
marker_pheatmap(geneset, seuratset, scale.max, scale.min, scale.threshold)
}
\arguments{
\item{geneset, seuratset, scale.max, scale.min, scale.threshold:}{input parameters}
}
\value{

}
\description{
This function draws heatmap with cluster-specific marker genes
}
\keyword{marker_pheatmap}
