% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat2harmony.R
\name{seurat2harmony}
\alias{seurat2harmony}
\title{seurat2harmony() Function}
\usage{
seurat2harmony(orig.seurat, dims, harmony.dims, batch.term, nfeatures)
}
\arguments{
\item{orig.seurat}{,dims, harmony.dims, batch.term, nfeatures}
}
\value{

}
\description{
This function operate harmony process with seurat object (as input data)
}
\keyword{seurat2harmony}
