% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_gene_ExpressionPlot.R
\name{seurat_gene_expression_map}
\alias{seurat_gene_expression_map}
\title{seurat_gene_expression_map() Function}
\usage{
seurat_gene_expression_map(seurat.obj, Gene, order, reduction_type, pt.size)
}
\arguments{
\item{seurat.obj, Gene, order, reduction_type, pt.size:}{input parameters for visualization}
}
\value{

}
\description{
This function draws gene expression on dimension reduction plot
}
\keyword{seurat_gene_expression_map}
