% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snRNAseq_CITEseq_Integration.R
\name{snRNAseq_CITEseq_Integration}
\alias{snRNAseq_CITEseq_Integration}
\title{snRNAseq_CITEseq_Integration() Function}
\usage{
snRNAseq_CITEseq_Integration(snRNAseq.dir, CITEseq.dir)
}
\arguments{
\item{snRNAseq.dir}{,CITEseq.dir : data path}
}
\value{

}
\description{
This function adds a and b
}
\keyword{snRNAseq_CITEseq_Integration}
